requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pymongo==4.8.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.8.2
python-dotenv==1.0.1
email-validator==2.2.0
requests==2.32.3